Include D3STATE indexes in init_consituent routines (changes in PERL script)
The routines of constituents initialization  were modified to include also
the passing of D3STATE indexes to check if indexes are zero.
This avoid the memory writing to the zero dimension of D3STATE that
generates a memory corruption using Intel 2015 compiler.
The PERL script generator and init_var_bfm.proto were modified.